requirements.txt
streamlit
pandas
numpy==1.26.4
scikit-surprise
